tests: treat empty string as falsey

+++ cdr-t5/build.sh
@@ -10,4 +10,6 @@ docker buildx build ${@:1} \
     --tag $IMAGE:$VERSION \
     --platform ${BUILDX_PLATFORM:-linux/arm64,linux/amd64} \
     .
-docker run --rm $IMAGE:$VERSION pytest
+if test -z $TESTS_SKIP_CPU; then
+  docker run --rm $IMAGE:$VERSION pytest
+fi
