+++ t2a-bark/build.sh
@@ -9,7 +9,7 @@ docker buildx build ${@:1} \
     --tag $IMAGE:latest \
     --tag $IMAGE:$VERSION \
     .
-if test -z ${TESTS_SKIP_GPU+x}; then
+if test -z $TESTS_SKIP_GPU; then
   docker run --rm --gpus all $IMAGE:$VERSION pytest
 fi
 
@@ -21,4 +21,6 @@ docker buildx build ${@:1} \
     --tag $IMAGE:$VERSION \
     --platform ${BUILDX_PLATFORM:-linux/arm64,linux/amd64} \
     .
-docker run --rm $IMAGE:$VERSION pytest
+if test -z $TESTS_SKIP_CPU; then
+  docker run --rm $IMAGE:$VERSION pytest
+fi
